moved mediaview to Sensus.UI namesace
