+++ Sensus.Shared/Sensus.Shared.projitems
@@ -33,7 +33,7 @@
     <Compile Include="$(MSBuildThisFileDirectory)MediaCacheModes.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MediaObject.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MediaStorageMethods.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)MediaView.cs" />
+    <Compile Include="$(MSBuildThisFileDirectory)UI\MediaView.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)MessageEventDatum.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Notifications\UserMessage.cs" />
     <Compile Include="$(MSBuildThisFileDirectory)Probes\Apps\ApplicationUsageEventProbe.cs" />
